+++ src/pages/tic-tac-toe/tic-tac-toe-single/tic-tac-toe-single.component.tsx
@@ -6,17 +6,21 @@ import GlobalButton from '../../../components/global-button/global-button.compon
 
 const TicTacToeSingle: FC = () => {
   return (
-    <div className='global-page-container'>
-      <div className='pixelated-grid'></div>
-      <div className='tic-tac-toe-single-container'>
-        <div className='tic-tac-toe-single-back-button'>
-          <GlobalButton label='<- Back' route='/tic-tac-toe-home' />
-        </div>
-        <header className='tic-tac-toe-single-header'>
-          <h1 className='single-player-header-text'>Single Player</h1>
-        </header>
-        <div className='tic-tac-toe-single-content'>
-          <p>Coming Soon...</p>
+    <div className='tic-tac-toe-single-container'>
+      <div className='tic-tac-toe-single-back-button'>
+        <GlobalButton label='<- Back' route='/tic-tac-toe-home' />
+      </div>
+      <header className='tic-tac-toe-single-header'>
+        <h1 className='single-player-header-text'>Single Player</h1>
+      </header>
+      <div className='tic-tac-toe-single-content'>
+        <p>Coming Soon...</p>
+        <div>
+          <p>
+            <span className='arrow-icon'>&#x2193;</span> Click Here for 2P{' '}
+            <span className='arrow-icon'>&#x2193;</span>
+          </p>
+          <GlobalButton label='Two Player' route='/tic-tac-toe-2p' />
         </div>
       </div>
     </div>